+++ src/ast.c
@@ -30,7 +30,6 @@ JL_DLLEXPORT jl_sym_t *jl_module_sym;
 JL_DLLEXPORT jl_sym_t *jl_slot_sym;
 JL_DLLEXPORT jl_sym_t *jl_export_sym;
 JL_DLLEXPORT jl_sym_t *jl_public_sym;
-JL_DLLEXPORT jl_sym_t *jl_import_sym;
 JL_DLLEXPORT jl_sym_t *jl_toplevel_sym;
 JL_DLLEXPORT jl_sym_t *jl_quote_sym;
 JL_DLLEXPORT jl_sym_t *jl_line_sym;
@@ -51,7 +50,6 @@ JL_DLLEXPORT jl_sym_t *jl_pop_exception_sym;
 JL_DLLEXPORT jl_sym_t *jl_exc_sym;
 JL_DLLEXPORT jl_sym_t *jl_error_sym;
 JL_DLLEXPORT jl_sym_t *jl_new_sym;
-JL_DLLEXPORT jl_sym_t *jl_using_sym;
 JL_DLLEXPORT jl_sym_t *jl_splatnew_sym;
 JL_DLLEXPORT jl_sym_t *jl_block_sym;
 JL_DLLEXPORT jl_sym_t *jl_new_opaque_closure_sym;
@@ -349,8 +347,6 @@ void jl_init_common_symbols(void)
     jl_module_sym = jl_symbol("module");
     jl_export_sym = jl_symbol("export");
     jl_public_sym = jl_symbol("public");
-    jl_import_sym = jl_symbol("import");
-    jl_using_sym = jl_symbol("using");
     jl_assign_sym = jl_symbol("=");
     jl_method_sym = jl_symbol("method");
     jl_exc_sym = jl_symbol("the_exception");
