+++ src/julia_internal.h
@@ -1849,7 +1849,6 @@ extern JL_DLLEXPORT jl_sym_t *jl_module_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_slot_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_export_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_public_sym;
-extern JL_DLLEXPORT jl_sym_t *jl_import_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_toplevel_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_quote_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_line_sym;
@@ -1871,7 +1870,6 @@ extern JL_DLLEXPORT jl_sym_t *jl_pop_exception_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_exc_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_error_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_new_sym;
-extern JL_DLLEXPORT jl_sym_t *jl_using_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_splatnew_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_block_sym;
 extern JL_DLLEXPORT jl_sym_t *jl_new_opaque_closure_sym;
