+++ src/toplevel.c
@@ -487,110 +487,10 @@ static void body_attributes(jl_array_t *body, int *has_ccall, int *has_defs, int
     *forced_compile = jl_has_meta(body, jl_force_compile_sym);
 }
 
-extern size_t jl_require_world;
-static jl_module_t *call_require(jl_task_t *ct, jl_module_t *mod, jl_sym_t *var) JL_GLOBALLY_ROOTED
-{
-    JL_TIMING(LOAD_IMAGE, LOAD_Require);
-    jl_timing_printf(JL_TIMING_DEFAULT_BLOCK, "%s", jl_symbol_name(var));
-
-    int build_mode = jl_options.incremental && jl_generating_output();
-    jl_module_t *m = NULL;
-    static jl_value_t *require_func = NULL;
-    if (require_func == NULL && jl_base_module != NULL) {
-        require_func = jl_get_global(jl_base_module, jl_symbol("require"));
-    }
-    if (require_func != NULL) {
-        ct->world_age = jl_atomic_load_acquire(&jl_world_counter);
-        if (build_mode && jl_require_world < ct->world_age)
-            ct->world_age = jl_require_world;
-        jl_value_t *reqargs[3];
-        reqargs[0] = require_func;
-        reqargs[1] = (jl_value_t*)mod;
-        reqargs[2] = (jl_value_t*)var;
-        m = (jl_module_t*)jl_apply(reqargs, 3);
-    }
-    if (m == NULL || !jl_is_module(m)) {
-        jl_errorf("failed to load module %s", jl_symbol_name(var));
-    }
-    ct->world_age = jl_atomic_load_acquire(&jl_world_counter);
-    return m;
-}
-
-// either:
-//   - sets *name and returns the module to import *name from
-//   - sets *name to NULL and returns a module to import
-// also updates world_age
-static jl_module_t *eval_import_path(jl_task_t *ct, jl_module_t *where, jl_module_t *from JL_PROPAGATES_ROOT,
-                                     jl_array_t *args, jl_sym_t **name, const char *keyword) JL_GLOBALLY_ROOTED
-{
-    if (jl_array_nrows(args) == 0)
-        jl_errorf("malformed \"%s\" statement", keyword);
-    jl_sym_t *var = (jl_sym_t*)jl_array_ptr_ref(args, 0);
-    size_t i = 1;
-    jl_module_t *m = NULL;
-    *name = NULL;
-    if (!jl_is_symbol(var))
-        jl_type_error(keyword, (jl_value_t*)jl_symbol_type, (jl_value_t*)var);
-
-    if (from != NULL) {
-        m = from;
-        i = 0;
-    }
-    else if (var != jl_dot_sym) {
-        // `A.B`: call the loader to obtain the root A in the current environment.
-        if (jl_core_module && var == jl_core_module->name) {
-            m = jl_core_module;
-        }
-        else if (jl_base_module && var == jl_base_module->name) {
-            m = jl_base_module;
-        }
-        else {
-            m = call_require(ct, where, var);
-        }
-        if (i == jl_array_nrows(args))
-            return m;
-    }
-    else {
-        // `.A.B.C`: strip off leading dots by following parent links
-        m = where;
-        while (1) {
-            if (i >= jl_array_nrows(args))
-                jl_error("invalid module path");
-            var = (jl_sym_t*)jl_array_ptr_ref(args, i);
-            if (var != jl_dot_sym)
-                break;
-            i++;
-            assert(m);
-            m = m->parent;
-        }
-    }
-
-    ct->world_age = jl_atomic_load_acquire(&jl_world_counter);
-
-    while (1) {
-        var = (jl_sym_t*)jl_array_ptr_ref(args, i);
-        if (!jl_is_symbol(var))
-            jl_type_error(keyword, (jl_value_t*)jl_symbol_type, (jl_value_t*)var);
-        if (var == jl_dot_sym)
-            jl_errorf("invalid %s path: \".\" in identifier path", keyword);
-        if (i == jl_array_nrows(args)-1)
-            break;
-        m = (jl_module_t*)jl_eval_global_var(m, var);
-        JL_GC_PROMISE_ROOTED(m);
-        if (!jl_is_module(m))
-            jl_errorf("invalid %s path: \"%s\" does not name a module", keyword, jl_symbol_name(var));
-        i++;
-    }
-    *name = var;
-    return m;
-}
-
 int jl_is_toplevel_only_expr(jl_value_t *e) JL_NOTSAFEPOINT
 {
     return jl_is_expr(e) &&
         (((jl_expr_t*)e)->head == jl_module_sym ||
-         ((jl_expr_t*)e)->head == jl_import_sym ||
-         ((jl_expr_t*)e)->head == jl_using_sym ||
          ((jl_expr_t*)e)->head == jl_export_sym ||
          ((jl_expr_t*)e)->head == jl_public_sym ||
          ((jl_expr_t*)e)->head == jl_thunk_sym ||
@@ -608,10 +508,9 @@ int jl_needs_lowering(jl_value_t *e) JL_NOTSAFEPOINT
         return 0;
     jl_expr_t *ex = (jl_expr_t*)e;
     jl_sym_t *head = ex->head;
-    if (head == jl_module_sym || head == jl_import_sym || head == jl_using_sym ||
-        head == jl_export_sym || head == jl_public_sym || head == jl_thunk_sym ||
-        head == jl_toplevel_sym || head == jl_error_sym || head == jl_incomplete_sym ||
-        head == jl_method_sym) {
+    if (head == jl_module_sym || head == jl_export_sym || head == jl_public_sym ||
+        head == jl_thunk_sym || head == jl_toplevel_sym || head == jl_error_sym ||
+        head == jl_incomplete_sym || head == jl_method_sym) {
         return 0;
     }
     if (head == jl_global_sym || head == jl_const_sym) {
@@ -649,62 +548,6 @@ JL_DLLEXPORT jl_method_instance_t *jl_method_instance_for_thunk(jl_code_info_t *
 
     JL_GC_POP();
     return mi;
-}
-
-static void import_module(jl_task_t *ct, jl_module_t *JL_NONNULL m, jl_module_t *import, jl_sym_t *asname)
-{
-    assert(m);
-    jl_sym_t *name = asname ? asname : import->name;
-    // TODO: this is a bit race-y with what error message we might print
-    jl_binding_t *b = jl_get_module_binding(m, name, 1);
-    jl_binding_partition_t *bpart = jl_get_binding_partition(b, ct->world_age);
-    enum jl_partition_kind kind = jl_binding_kind(bpart);
-    if (!jl_bkind_is_some_implicit(kind) && kind != PARTITION_KIND_DECLARED) {
-        // Unlike regular constant declaration, we allow this as long as we eventually end up at a constant.
-        jl_walk_binding_inplace(&b, &bpart, ct->world_age);
-        if (jl_bkind_is_some_constant(jl_binding_kind(bpart))) {
-            // Already declared (e.g. on another thread) or imported.
-            if (bpart->restriction == (jl_value_t*)import)
-                return;
-        }
-        jl_errorf("importing %s into %s conflicts with an existing global",
-                    jl_symbol_name(name), jl_symbol_name(m->name));
-    }
-    jl_declare_constant_val2(b, m, name, (jl_value_t*)import, PARTITION_KIND_CONST_IMPORT);
-}
-
-// in `import A.B: x, y, ...`, evaluate the `A.B` part if it exists
-static jl_module_t *eval_import_from(jl_task_t *ct, jl_module_t *m JL_PROPAGATES_ROOT, jl_expr_t *ex, const char *keyword)
-{
-    if (jl_expr_nargs(ex) == 1 && jl_is_expr(jl_exprarg(ex, 0))) {
-        jl_expr_t *fr = (jl_expr_t*)jl_exprarg(ex, 0);
-        if (fr->head == jl_colon_sym) {
-            if (jl_expr_nargs(fr) > 0 && jl_is_expr(jl_exprarg(fr, 0))) {
-                jl_expr_t *path = (jl_expr_t*)jl_exprarg(fr, 0);
-                if (((jl_expr_t*)path)->head == jl_dot_sym) {
-                    jl_sym_t *name = NULL;
-                    jl_module_t *from = eval_import_path(ct, m, NULL, path->args, &name, keyword);
-                    if (name != NULL) {
-                        from = (jl_module_t*)jl_eval_global_var(from, name);
-                        if (!from || !jl_is_module(from))
-                            jl_errorf("invalid %s path: \"%s\" does not name a module", keyword, jl_symbol_name(name));
-                    }
-                    return from;
-                }
-            }
-            jl_errorf("malformed \"%s:\" statement", keyword);
-        }
-    }
-    return NULL;
-}
-
-static void check_macro_rename(jl_sym_t *from, jl_sym_t *to, const char *keyword)
-{
-    char *n1 = jl_symbol_name(from), *n2 = jl_symbol_name(to);
-    if (n1[0] == '@' && n2[0] != '@')
-        jl_errorf("cannot rename macro \"%s\" to non-macro \"%s\" in \"%s\"", n1, n2, keyword);
-    if (n1[0] != '@' && n2[0] == '@')
-        jl_errorf("cannot rename non-macro \"%s\" to macro \"%s\" in \"%s\"", n1, n2, keyword);
 }
 
 // Eval `throw(ErrorException(msg)))` in module `m`.
@@ -826,111 +669,6 @@ JL_DLLEXPORT jl_value_t *jl_toplevel_eval_flex(jl_module_t *JL_NONNULL m, jl_val
         jl_value_t *val = jl_eval_module_expr(m, ex);
         JL_GC_POP();
         return val;
-    }
-    else if (head == jl_using_sym) {
-        jl_sym_t *name = NULL;
-        jl_module_t *from = eval_import_from(ct, m, ex, "using");
-        size_t i = 0;
-        if (from) {
-            i = 1;
-            ex = (jl_expr_t*)jl_exprarg(ex, 0);
-        }
-        for (; i < jl_expr_nargs(ex); i++) {
-            jl_value_t *a = jl_exprarg(ex, i);
-            if (jl_is_expr(a) && ((jl_expr_t*)a)->head == jl_dot_sym) {
-                name = NULL;
-                jl_module_t *import = eval_import_path(ct, m, from, ((jl_expr_t*)a)->args, &name, "using");
-                if (from) {
-                    // `using A: B` and `using A: B.c` syntax
-                    jl_module_use(ct, m, import, name);
-                }
-                else {
-                    jl_module_t *u = import;
-                    if (name != NULL)
-                        u = (jl_module_t*)jl_eval_global_var(import, name);
-                    if (!jl_is_module(u))
-                        jl_eval_errorf(m, *toplevel_filename, *toplevel_lineno,
-                            "invalid using path: \"%s\" does not name a module",
-                                       jl_symbol_name(name));
-                    // `using A` and `using A.B` syntax
-                    jl_module_using(m, u);
-                    if (m == jl_main_module && name == NULL) {
-                        // TODO: for now, `using A` in Main also creates an explicit binding for `A`
-                        // This will possibly be extended to all modules.
-                        import_module(ct, m, u, NULL);
-                    }
-                }
-                continue;
-            }
-            else if (from && jl_is_expr(a) && ((jl_expr_t*)a)->head == jl_as_sym && jl_expr_nargs(a) == 2 &&
-                     jl_is_expr(jl_exprarg(a, 0)) && ((jl_expr_t*)jl_exprarg(a, 0))->head == jl_dot_sym) {
-                jl_sym_t *asname = (jl_sym_t*)jl_exprarg(a, 1);
-                if (jl_is_symbol(asname)) {
-                    jl_expr_t *path = (jl_expr_t*)jl_exprarg(a, 0);
-                    name = NULL;
-                    jl_module_t *import = eval_import_path(ct, m, from, ((jl_expr_t*)path)->args, &name, "using");
-                    assert(name);
-                    check_macro_rename(name, asname, "using");
-                    // `using A: B as C` syntax
-                    jl_module_use_as(ct, m, import, name, asname);
-                    continue;
-                }
-            }
-            jl_eval_errorf(m, *toplevel_filename, *toplevel_lineno,
-                "syntax: malformed \"using\" statement");
-        }
-        JL_GC_POP();
-        ct->world_age = last_age;
-        return jl_nothing;
-    }
-    else if (head == jl_import_sym) {
-        jl_sym_t *name = NULL;
-        jl_module_t *from = eval_import_from(ct, m, ex, "import");
-        size_t i = 0;
-        if (from) {
-            i = 1;
-            ex = (jl_expr_t*)jl_exprarg(ex, 0);
-        }
-        for (; i < jl_expr_nargs(ex); i++) {
-            jl_value_t *a = jl_exprarg(ex, i);
-            if (jl_is_expr(a) && ((jl_expr_t*)a)->head == jl_dot_sym) {
-                name = NULL;
-                jl_module_t *import = eval_import_path(ct, m, from, ((jl_expr_t*)a)->args, &name, "import");
-                if (name == NULL) {
-                    // `import A` syntax
-                    import_module(ct, m, import, NULL);
-                }
-                else {
-                    // `import A.B` or `import A: B` syntax
-                    jl_module_import(ct, m, import, name);
-                }
-                continue;
-            }
-            else if (jl_is_expr(a) && ((jl_expr_t*)a)->head == jl_as_sym && jl_expr_nargs(a) == 2 &&
-                     jl_is_expr(jl_exprarg(a, 0)) && ((jl_expr_t*)jl_exprarg(a, 0))->head == jl_dot_sym) {
-                jl_sym_t *asname = (jl_sym_t*)jl_exprarg(a, 1);
-                if (jl_is_symbol(asname)) {
-                    jl_expr_t *path = (jl_expr_t*)jl_exprarg(a, 0);
-                    name = NULL;
-                    jl_module_t *import = eval_import_path(ct, m, from, ((jl_expr_t*)path)->args, &name, "import");
-                    if (name == NULL) {
-                        // `import A as B` syntax
-                        import_module(ct, m, import, asname);
-                    }
-                    else {
-                        check_macro_rename(name, asname, "import");
-                        // `import A.B as C` syntax
-                        jl_module_import_as(ct, m, import, name, asname);
-                    }
-                    continue;
-                }
-            }
-            jl_eval_errorf(m, *toplevel_filename, *toplevel_lineno,
-                "syntax: malformed \"import\" statement");
-        }
-        JL_GC_POP();
-        ct->world_age = last_age;
-        return jl_nothing;
     }
     else if (head == jl_export_sym || head == jl_public_sym) {
         int exp = (head == jl_export_sym);
